microservice-parser-generation/requirements.txt
kafka-python==2.0.2
configparser==5.2.0
google-cloud-storage==2.4.0
confluent-kafka==1.9.0